+++ src/test/java/com/caoccao/javet/values/reference/builtin/TestV8ValueBuiltInJson.java
@@ -1,3 +1,20 @@
+/*
+ *   Copyright (c) 2021. caoccao.com Sam Cao
+ *   All rights reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
 package com.caoccao.javet.values.reference.builtin;
 
 import com.caoccao.javet.BaseTestJavetRuntime;
@@ -18,7 +35,7 @@ public class TestV8ValueBuiltInJson extends BaseTestJavetRuntime {
         v8Runtime.getExecutor(codeString).executeVoid();
         try (V8ValueObject v8ValueObject = v8Runtime.getGlobalObject().get("b")) {
             assertEquals(2, v8ValueObject.getInteger("z"));
-            try (V8ValueBuiltInJson v8ValueBuiltInJson = v8Runtime.getGlobalObject().getJson()) {
+            try (V8ValueBuiltInJson v8ValueBuiltInJson = v8Runtime.getGlobalObject().getBuiltInJson()) {
                 v8ValueBuiltInJson.stringify(v8ValueObject);
             } catch (JavetExecutionException e) {
                 assertEquals(